Constraints class

R6 class with constraints.

The constraints property on Table Schema Fields can be used by consumers to list constraints for validating field values. For example, validating the data in a Tabular Data Resource against its Table Schema; or as a means to validate data being collected or updated via a data entry interface.

All constraints MUST be tested against the logical representation of data, and the physical representation of constraint values MAY be primitive types as possible in JSON, or represented as strings that are castable with the type and format rules of the field.

Format

R6Class object.

Value

Object of R6Class .

See also

Methods

Public methods

Constraints$clone()

Method clone()

The objects of this class are cloneable with this method.

Usage

Constraints$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.