Readable class

Readable class that allows typed access to its members

Format

R6Class object.

Value

Object of R6Class.

Methods

Public methods

Readable$new()

Readable$read()

Readable$pipe()

Readable$unpipe()

Readable$pause()

Readable$resume()

Readable$setEncoding()

Readable$isPaused()

Readable$unshift()

Readable$destroy()

Readable$push()

Readable$onData()

Readable$on.close()

Readable$on.data()

Readable$on.end()

Readable$on.error()

Readable$on.readable()

Readable$clone()

Method new()

Usage

Readable$new(options = list())

Method read()

Usage

Readable$read(size = NULL)

Method pipe()

Usage

Readable$pipe(destination, options = list())

Method unpipe()

Usage

Readable$unpipe(destination)

Method pause()

Usage

Readable$pause()

Method resume()

Usage

Readable$resume()

Method setEncoding()

Usage

Readable$setEncoding()

Method isPaused()

Usage

Readable$isPaused()

Method unshift()

Usage

Readable$unshift(chunk)

Method destroy()

Usage

Readable$destroy()

Method push()

Usage

Readable$push(chunk, encoding)

Method onData()

Usage

Readable$onData(chunk)

Method on.close()

Usage

Readable$on.close(handler, unsubscribe = FALSE)

Method on.data()

Usage

Readable$on.data(handler, unsubscribe = FALSE)

Method on.end()

Usage

Readable$on.end(handler, unsubscribe = FALSE)

Method on.error()

Usage

Readable$on.error(handler, unsubscribe = FALSE)

Method on.readable()

Usage

Readable$on.readable(handler, unsubscribe = FALSE)

Method clone()

The objects of this class are cloneable with this method.

Usage

Readable$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.