Types class

R6 class with Types and Formats.

type and format properties are used to give the type of the field (string, number etc) - see types and formats for more details. If type is not provided a consumer should assume a type of "string".

A field's type property is a string indicating the type of this field.

A field's format property is a string, indicating a format for the field type.

Both type and format are optional: in a field descriptor, the absence of a type property indicates that the field is of the type "string", and the absence of a format property indicates that the field's type format is "default".

Types are based on the type set of json-schema with some additions and minor modifications (cf other type lists include those in Elasticsearch types).

Format

R6Class object.

Value

Object of R6Class .

See also

Methods

Public methods

Types$clone()

Method clone()

The objects of this class are cloneable with this method.

Usage

Types$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.